+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,16 @@
+package duke.command;
+
+import duke.exception.InvalidArgumentException;
+import duke.parser.Action;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.TaskList;
+import duke.task.Todo;
+import duke.ui.Messages;
+import duke.ui.Ui;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
