+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class ExitCommand extends Command {
 
     public void execute(TaskList tasks, Ui ui, Storage storage) {
