+++ src/main/java/duke/exception/OutOfBoundsException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class OutOfBoundsException extends DukeException {
     public OutOfBoundsException(String message) {
         super(message);
