+++ src/main/java/duke/exception/UnknownException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class UnknownException extends DukeException {
 
     public UnknownException(String message) {
