+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,7 @@
+package duke.storage;
+
+import duke.task.*;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
