+++ src/main/java/duke/ui/Messages.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 public class Messages {
     public static final String LINE_SEPARATOR = System.lineSeparator();
